Added CMakeLists.txt to Server Module

+++ server/Server.h
@@ -7,7 +7,7 @@
 
 #include "../ServerSocket.h"
 #include "../Socket.h"
-#include "../HttpWorkerThread.h";
+#include "../HttpWorkerThread.h"
 
 #define TIMEOUT 30
 #define SERVER_NAME "dummy"
@@ -15,17 +15,17 @@
 
 class Server {
     public:
-        Server(std::string &port_number, int backlog, int max_workers);
+        Server(std::string &port_number, unsigned short backlog, unsigned long max_workers);
         ~Server();
         void run();
         std::string get_port_number() const;
-        int get_backlog() const;
-        int get_max_workers() const;
+        unsigned short get_backlog() const;
+        unsigned long get_max_workers() const;
     private:
         ServerSocket *server_socket;
         std::string port_number;
-        int backlog;
-        int max_workers;
+        unsigned short backlog;
+        unsigned long max_workers;
         std::vector<HttpWorkerThread *> workers_pool;
 };
